tests done, bear & wolf fixed, changes of files
Tests:
W1: 100%
W2: 100%
W3: 100%

Bear:
If the bear doesn't have a territory it will create a territory at a random location without a blocking object on it. This change will throw an exception in case all tiles has nonblocking objects on it.

Wolf:
Fixed a case where the wolf would move to a nearest member that didn't exist